services/crawler/requirements.txt
requests
beautifulsoup4
selenium
webdriver_manager
supabase
python-dotenv